BASH/PHP Extras
the BASH script creates a file unique to the record you want to update
and uses the contents of that file in CloudFlare updates. if no file
exists, it gets all the records from CloudFlare and sends them to your
web service for processing, then creates the unique file.  started PHP
to process file creation if a new record is needed

also cleaned up BASH if/then/else control structure